CORE-3780 move config helper to config core module, add comment regarding removal of boot config keys, rename config exception and add new one.
