+++ libs/configuration/configuration-core/src/main/java/net/corda/libs/configuration/helper/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.libs.configuration.helper;
+
+import org.osgi.annotation.bundle.Export;
